fixed canvas + HUD

+++ Assets/Scripts/HealthMeter.cs
@@ -2,7 +2,7 @@ using UnityEngine;
 
 public class HealthMeter : Meter
 {
-    public void UpdateHealthMeter(int value)
+    public void UpdateHealthMeter(float value)
     {
         fillBar.fillAmount = value / 100;
     }
